refactor: dry out error pages
