+++ containers/ecr-viewer/src/app/view-data/components/SideNav.tsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import { SideNav as UswdsSideNav } from "@trussworks/react-uswds";
 import { toKebabCase } from "@/app/utils/format-utils";
 import classNames from "classnames";
-import { BackButton } from "./BackButton";
+import { BackButton } from "../../components/BackButton";
 import { env } from "next-runtime-env";
 import { SideNavLoadingSkeleton } from "./LoadingComponent";
